backend/requirements.txt
django
djangorestframework
django-cors-headers  # To handle CORS for frontend
